requirements.txt
requests
Pillow
pytesseract
beautifulsoup4